Don't crash if README_STARTUP is missing
On Windows inside a py2exe'd environment, README_STARTUP does not
get copied into library.zip. Added a warning and a skip if
the file IPython/config/profile/README_STARTUP is missing.

+++ IPython/core/profiledir.py
@@ -102,7 +102,11 @@ class ProfileDir(LoggingConfigurable):
             os.mkdir(self.startup_dir)
         readme = os.path.join(self.startup_dir, 'README')
         src = os.path.join(get_ipython_package_dir(), u'config', u'profile', u'README_STARTUP')
-        if not os.path.exists(readme):
+
+        if not os.path.exists(src):
+            self.log.warn("Could not copy README_STARTUP to startup dir. Source file %s does not exist." % src)
+
+        if os.path.exists(src) and not os.path.exists(readme):
             shutil.copy(src, readme)
 
     def _security_dir_changed(self, name, old, new):
